extracted `FwdAnalysis` from `astutils.{cpp|h}`
