+++ lib/cppcheck.vcxproj.filters
@@ -62,6 +62,9 @@
     <ClCompile Include="errortypes.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
+    <ClCompile Include="fwdanalysis.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
     <ClCompile Include="mathlib.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
@@ -230,6 +233,9 @@
       <Filter>Header Files</Filter>
     </ClInclude>
     <ClInclude Include="errorlogger.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="fwdanalysis.h">
       <Filter>Header Files</Filter>
     </ClInclude>
     <ClInclude Include="mathlib.h">
